+++ anxcloud/provider.go
@@ -74,6 +74,7 @@ func Provider(version string) *schema.Provider {
 			"anxcloud_dns_records":           dataSourceDNSRecords(),
 			"anxcloud_dns_zones":             datasourceDNSZones(),
 			"anxcloud_kubernetes_cluster":    dataSourceKubernetesCluster(),
+			"anxcloud_kubernetes_kubeconfig": dataSourceKubernetesKubeconfig(),
 			// Object Storage data sources
 			"anxcloud_object_storage_endpoints": dataSourceObjectStorageEndpoints(),
 			"anxcloud_object_storage_backends":  dataSourceObjectStorageBackends(),