+++ tests/e2e/kubernetes/versions.tf
@@ -6,14 +6,6 @@ terraform {
       source  = "hashicorp.com/anexia-it/anxcloud"
       version = "0.11.1"
     }
-    helm = {
-      source  = "hashicorp/helm"
-      version = "3.2.0"
-    }
-    kubernetes = {
-      source  = "hashicorp/kubernetes"
-      version = "3.2.1"
-    }
   }
 }